Created a reusable details modal and implemented it on jobMonitoring. Removed the unnecessary MonitoringDetailsModal.jsx inside jobMonitoring.
